+++ src/main/kotlin/site/siredvin/peripheralworks/PeripheralWorks.kt
@@ -20,6 +20,7 @@ object PeripheralWorks: ModInitializer {
         ComputerCraftAPI.registerPeripheralProvider(ComputerCraftProxy::peripheralProvider)
         // Load all integrations
         Platform.maybeLoadIntegration("team_reborn_energy").ifPresent { (it as Runnable).run() }
+        Platform.maybeLoadIntegration("lifts").ifPresent { (it as Runnable).run() }
         // Register configuration
         // Pretty important to setup configuration after integration loading!
         ModLoadingContext.registerConfig(MOD_ID, ModConfig.Type.COMMON, ConfigHolder.COMMON_SPEC)